Remove bulk data before re-adding it
This is relevant because inserting a list may have an implicit deletion:

BEFORE:
   house, tree, road
AFTER
   house, office, tree

This change implicitly deletes the 'road' and adds an 'office'. I could
calculate the delta to hit the DB as little as possible but I'm too
lazy for that, so instead I DELETE the old list and INSERT the new one.